requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
faiss-cpu
numpy
pandas
pypdf